requirements.txt
python-telegram-bot==21.3
replicate>=0.34.0
Pillow>=10.4.0
python-dotenv>=1.0.0
